// apps/pos/src/app/views/SelfServe/types.ts
export interface OrderCreateResponse extends CreateOrderPayload {
  amount: {
    base: number;
    gst: number;
    total: number;
  };
  rental_amount: {
    base: number;
    gst: number;
    total: number;
  } | null;
}

export type OrderObjectType = {
  id: string;
  amount: number;
  amount_paid: number;
  amount_due: number;
  currency: string;
  status: string;
  created_at: number;
};

export type RefundStatusTypes = 'failed' | 'processing' | 'processed';
export interface OrderDetailsItem extends OrderCreateResponse {
  id: string;
  order_id: string;
  merchant_id: string;
  created_at: number;
  arriving_at: number;
  delivered_at: number;
  rejected_at: number;
  rejection_reasons: {
    error_code: string;
    error_description: string;
    error_reason: string;
  } | null;
  status: 'paid' | 'delivered' | 'rejected';
  order?: OrderObjectType | null;
  sales_code?: string;
  payment?: {
    amount: number;
    status: string;
    refund_status: null | string;
    captured: boolean;
    created_at: number;
  } | null;
  refund?: {
    id: string;
    amount: number;
    payment_id: string;
    created_at: number;
    status: RefundStatusTypes;
    acquirer_data: Record<string, string>;
  } | null;
}

export type OrderStatusTypes =
  | 'ORDER_RECEIVED'
  | 'ORDER_DELIVERED'
  | 'REFUND_PENDING'
  | 'REFUND_INITIATED'
  | 'REFUND_COMPLETED'
  | 'ORDER_REJECTED'
  | 'ORDER_CONFIRMED';

export type PincodeInfo = {
  state: string;
  city: string;
  state_code: string;
};
// ...
export type CreateOrderPayload = {
  delivery_address: {
    name: string;
    address: string;
    city: string;
    country: string;
    state: string;
    pin_code: string;
    phone_no: string;
  };
  items: {
    code: string;
    count: number;
    period: string;
  }[];
};

export type ProcessedRefundObj = {
  id: string;
  refId: string;
  amount: number;
  status: 'processed' | 'pending' | 'failed';
};

export type DetailedPricingModel = {
  title: string | null;
  banner?: 'offer' | 'info';
  rows: {
    name: string;
    value: string | number;
    prevValue?: string | number;
    text?: string;
    isOfferOnlyField?: boolean;
  }[];
};

export interface ApiResponse<T> {
  status_code: number;
  success: boolean;
  data?: T;
  errors?: string[];
}

export type PlanType = {
  one_time_charge: number;
  plan_name: PricingTypes;
  rental_charges: number;
  setup_fee: number;
};

export type RateConfig = {
  advanced_rental_periods: number;
  rental_discount_periods: number;
  paper_roll_charges: number;
  plans: PlanType[];
};

export type DeviceMetaData = {
  charge_collection_product_id: string;
  rate_config_v2?: RateConfig | null;
};

export type ProductPricingMap = {
  name: string;
  code: string;
  entity_type?: string;
  rate_config: {
    monthly: number;
    lifetime: number;
    setup_fee: number;
  };
  metadata?: DeviceMetaData;
}[];
// ...
export type CommsOrderItem = {
  id: string;
  amount: {
    total: number;
  };
  refund: {
    amount: number;
    status: string;
  } | null;
};
// ...
export type PosActivationStatusTypes =
  | 'activated'
  | 'under_review'
  | 'needs_clarification'
  | 'rejected'
  | 'kyc_qualified_stb'
  | 'kyc_qualified_unactivated'
  | null;
// ...
export type PosAgreementSignIds = {
  merchantId: string;
  tncId: string;
  privacyId: string;
  pricingId?: string;
  agreement_consented_at_field: number;
};

type ConsentField = {
  type: string;
  templateId: string;
};

export type PosAgreementSignPayload = {
  terms_and_conditions_consent_field: ConsentField;
  privacy_consent_field: ConsentField;
  pricing_consent_field?: ConsentField;
  agreement_consented_at_field: number;
};

export type WorkflowData = {
  id: string;
  milestones: [ModularOnboardingMilestone];
  progress: number;
  status: string;
};

export type ModularOnboardingMilestone = {
  name: string;
  status: string;
  can_submit: boolean;
  steps: [ModularOnboardingStep];
  progress: number;
  meta: { template: string; title: string };
};
// ...
export type WorkflowConfig = {
  success: boolean;
  workflow_data: WorkflowData;
  onboarding_state: {
    components: string[];
    milestones: string[];
    steps: string[];
  };
  onboarding_status: string;
  country_code: string;
  onboarding_type: string;
  merchant_type: string;
};
// ...
export type PaginationParamsType = { skip: number | string; count: number | string };